Modify hash to use a single multiplication instead of using the murmurhash
