+++ test/sql/aggregate/test_aggregate.cpp
@@ -126,9 +126,10 @@ TEST_CASE("Test STRING_AGG operator", "[aggregate]") {
 	REQUIRE(CHECK_COLUMN(result, 0, {"a,b,i,j,p,x,y,z"}));
 	REQUIRE(CHECK_COLUMN(result, 1, {"a-b/i+j/p/x-y+z"}));
 
-	result = con.Query("SELECT STRING_AGG(x,','), STRING_AGG(x,y) FROM strings GROUP BY g");
-	REQUIRE(CHECK_COLUMN(result, 0, {"a,b", "x,y,z", "i,j", "p"}));
-	REQUIRE(CHECK_COLUMN(result, 1, {"a-b", "x-y+z", "i+j", "p"}));
+	result = con.Query("SELECT g, STRING_AGG(x,','), STRING_AGG(x,y) FROM strings GROUP BY g ORDER BY g");
+	REQUIRE(CHECK_COLUMN(result, 0, {1, 2, 3, 4}));
+	REQUIRE(CHECK_COLUMN(result, 1, {"a,b", "i,j", "p", "x,y,z"}));
+	REQUIRE(CHECK_COLUMN(result, 2, {"a-b", "i+j", "p", "x-y+z"}));
 
 	// test average on empty set
 	result = con.Query("SELECT STRING_AGG(x,','), STRING_AGG(x,y) FROM strings WHERE g > 100");
